backend down

+++ backend/src/resolvers/Mutation.js
@@ -2,13 +2,31 @@ const bcrybt=require('bcryptjs')
 const jwt=require('jsonwebtoken')
 const {randomBytes}=require('crypto')
 const {promisify}=require('util')
+const { hasPermission } = require('../utils');
 const Mutations = {
     async createItem(parent,args,ctx,info){
         //this return promise
-        const item=await ctx.db.mutation.createItem({data:{
-            ...args
-        }},info)
-        console.log(item)
+        if (!ctx.request.userId) {
+          throw new Error('You must be logged in to do that!');
+        }
+    
+        const item = await ctx.db.mutation.createItem(
+          {
+            data: {
+              // This is how to create a relationship between the Item and the User
+              user: {
+                connect: {
+                  id: ctx.request.userId,
+                },
+              },
+              ...args,
+            },
+          },
+          info
+        );
+    
+        //console.log(item);
+    
         return item;
     },
     updateItem(parent, args, ctx, info) {
@@ -27,10 +45,21 @@ const Mutations = {
           info
         );
       },
-      async deleteItem(parent, args, ctx, info){
+    async deleteItem(parent, args, ctx, info){
         const where= {id:args.id}
-        const item=await ctx.db.query.item({ where },`{ id title }`)
-        return ctx.db.mutation.deleteItem({ where},info)
+        const item = await ctx.db.query.item({ where }, `{ id title user { id }}`);
+    // 2. Check if they own that item, or have the permissions
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions.some(permission =>
+      ['ADMIN', 'ITEMDELETE'].includes(permission)
+    );
+
+    if (!ownsItem && !hasPermissions) {
+      throw new Error("You don't have permission to do that!");
+    }
+
+    // 3. Delete it!
+    return ctx.db.mutation.deleteItem({ where }, info);
       },
       async signup(parent, args, ctx, info){
         args.email=args.email.toLowerCase()
@@ -137,8 +166,162 @@ const Mutations = {
     });
     // 8. return the new user
     return updatedUser;
+      },
+    //new things need more reading
+    
+    
+    async updatePermissions(parent, args, ctx, info) {
+      // 1. Check if they are logged in
+      if (!ctx.request.userId) {
+        throw new Error('You must be logged in!');
       }
-      
+      // 2. Query the current user
+      const currentUser = await ctx.db.query.user(
+        {
+          where: {
+            id: ctx.request.userId,
+          },
+        },
+        info
+      );
+      // 3. Check if they have permissions to do this
+      hasPermission(currentUser, ['ADMIN', 'PERMISSIONUPDATE']);
+      // 4. Update the permissions
+      return ctx.db.mutation.updateUser(
+        {
+          data: {
+            permissions: {
+              set: args.permissions,
+            },
+          },
+          where: {
+            id: args.userId,
+          },
+        },
+        info
+      );
+    },
+    async addToCart(parent, args, ctx, info) {
+      // 1. Make sure they are signed in
+      const { userId } = ctx.request;
+      if (!userId) {
+        throw new Error('You must be signed in soooon');
+      }
+      // 2. Query the users current cart
+      const [existingCartItem] = await ctx.db.query.cartItems({
+        where: {
+          user: { id: userId },
+          item: { id: args.id },
+        },
+      });
+      // 3. Check if that item is already in their cart and increment by 1 if it is
+      if (existingCartItem) {
+        console.log('This item is already in their cart');
+        return ctx.db.mutation.updateCartItem(
+          {
+            where: { id: existingCartItem.id },
+            data: { quantity: existingCartItem.quantity + 1 },
+          },
+          info
+        );
+      }
+      // 4. If its not, create a fresh CartItem for that user!
+      return ctx.db.mutation.createCartItem(
+        {
+          data: {
+            user: {
+              connect: { id: userId },
+            },
+            item: {
+              connect: { id: args.id },
+            },
+          },
+        },
+        info
+      );
+    },
+    async removeFromCart(parent, args, ctx, info) {
+      // 1. Find the cart item
+      const cartItem = await ctx.db.query.cartItem(
+        {
+          where: {
+            id: args.id,
+          },
+        },
+        `{ id, user { id }}`
+      );
+      // 1.5 Make sure we found an item
+      if (!cartItem) throw new Error('No CartItem Found!');
+      // 2. Make sure they own that cart item
+      if (cartItem.user.id !== ctx.request.userId) {
+        throw new Error('Cheatin huhhhh');
+      }
+      // 3. Delete that cart item
+      return ctx.db.mutation.deleteCartItem(
+        {
+          where: { id: args.id },
+        },
+        info
+      );
+    },
+    async createOrder(parent, args, ctx, info) {
+      // 1. Query the current user and make sure they are signed in
+      const { userId } = ctx.request;
+      if (!userId) throw new Error('You must be signed in to complete this order.');
+      const user = await ctx.db.query.user(
+        { where: { id: userId } },
+        `{
+        id
+        name
+        email
+        cart {
+          id
+          quantity
+          item { title price id description image largeImage }
+        }}`
+      );
+      // 2. recalculate the total for the price
+      const amount = user.cart.reduce(
+        (tally, cartItem) => tally + cartItem.item.price * cartItem.quantity,
+        0
+      );
+      console.log(`Going to charge for a total of ${amount}`);
+      // 3. Create the stripe charge (turn token into $$$)
+      const charge = await stripe.charges.create({
+        amount,
+        currency: 'USD',
+        source: args.token,
+      });
+      // 4. Convert the CartItems to OrderItems
+      const orderItems = user.cart.map(cartItem => {
+        const orderItem = {
+          ...cartItem.item,
+          quantity: cartItem.quantity,
+          user: { connect: { id: userId } },
+        };
+        delete orderItem.id;
+        return orderItem;
+      });
+  
+      // 5. create the Order
+      const order = await ctx.db.mutation.createOrder({
+        data: {
+          total: charge.amount,
+          charge: charge.id,
+          items: { create: orderItems },
+          user: { connect: { id: userId } },
+        },
+      });
+      // 6. Clean up - clear the users cart, delete cartItems
+      const cartItemIds = user.cart.map(cartItem => cartItem.id);
+      await ctx.db.mutation.deleteManyCartItems({
+        where: {
+          id_in: cartItemIds,
+        },
+      });
+      // 7. Return the Order to the client
+      return order;
+    },
 };
 
 module.exports = Mutations;
